requirements.txt
ultralytics>=8.0.0
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
scipy>=1.10.0
filterpy>=1.4.5
websockets>=11.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
